fix: precision loss in SoundEffectPacket coordinates

+++ src/main/java/net/minestom/server/adventure/AdventurePacketConvertor.java
@@ -113,7 +113,7 @@ public class AdventurePacketConvertor {
         if (minestomSound == null) minestomSound = SoundEvent.of(sound.name(), null);
 
         final long seed = sound.seed().orElse(ThreadLocalRandom.current().nextLong());
-        return new SoundEffectPacket(minestomSound, sound.source(), (int) x, (int) y, (int) z, sound.volume(), sound.pitch(), seed);
+        return new SoundEffectPacket(minestomSound, sound.source(), x, y, z, sound.volume(), sound.pitch(), seed);
     }
 
     /**
